+++ ChangeLog.md
@@ -9,6 +9,7 @@ ChangeLog
 * #119: Support for detecting 'significant changes'.
 * #120: Support for `SCHEDULE-FORCE-SEND`.
 * #121: iCal demands parameters contain the + sign to be quoted.
+* #122: Don't generate REPLY messages for events that have been cancelled.
 
 3.3.1 (2014-08-18)
 ------------------